add additional tests for CDEs
cde_behavior_test is the start of a test suite for ML behavioral
testing. Right now, the only test it contains checks that the training
loss goes down over time for each CDE.

cde_regression_test compares the output of the CDE to a saved, "correct"
result. Since the results of the CDEs are pretty variable, this test
calculates the mean for each feature from the CDE output, and checks
that they are equal between arrays within a 15% relative tolerance (see numpy
allclose for details on what that means).